pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "avtomatika"
version = "1.0a2"
description = "A state-machine based orchestrator for long-running jobs."
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "aiohttp~=3.12",
    "aiocache~=0.12",
    "python-json-logger~=4.0",
    "graphviz~=0.21",
    "zstandard~=0.24",
    "aioprometheus~=23.12",
]

[project.optional-dependencies]
redis = ["redis~=6.4", "orjson~=3.11"]
history = ["aiosqlite~=0.21", "asyncpg~=0.30", "orjson~=3.11"]
telemetry = [
    "opentelemetry-api~=1.38",
    "opentelemetry-sdk~=1.38",
    "opentelemetry-exporter-otlp~=1.36",
    "opentelemetry-instrumentation-aiohttp-client~=0.59b0",
]
test = [
    "pytest~=8.4",
    "pytest-asyncio~=1.1",
    "fakeredis~=2.31",
    "pytest-aiohttp~=1.1",
    "pytest-mock~=3.14",
    "aioresponses~=0.7",
    "backports.zstd",
    "opentelemetry-instrumentation-aiohttp-client",
]
all = [
    "avtomatika[redis]",
    "avtomatika[history]",
    "avtomatika[telemetry]",
]

[project.urls]
"Homepage" = "https://github.com/avtomatika-ai/avtomatika"
"Bug Tracker" = "https://github.com/avtomatika-ai/avtomatika/issues"

[tool.setuptools.package-data]
"avtomatika" = ["api.html"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
markers = [
    "e2e: marks tests as end-to-end tests",
]

[tool.ruff]
src = ["src", "tests"]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "A", "SIM", "C4", "PLC2401"]

[tool.ruff.lint.isort]
known-first-party = ["avtomatika"]

[tool.ruff.format]
quote-style = "double"
skip-magic-trailing-comma = false

[tool.mypy]
python_version = "3.11"
packages = ["avtomatika"]
package_path = "src"
disallow_incomplete_defs = true
disallow_untyped_calls = true
no_implicit_optional = true
ignore_missing_imports = true
follow_imports = "skip"
warn_unused_ignores = true
warn_return_any = true
disallow_any_unimported = false
exclude = ["tests/"]
